page구현

+++ app/src/main/java/com/wafflestudio/snugo/models/Route.kt
@@ -4,6 +4,6 @@ data class Route(
     val id: String,
     val buildings: List<Building>,
     val count: Long,
-    val averagePathLength: Long,
-    val averageTime: Long,
+    val averagePathLength: Double,
+    val averageTime: Double,
 )
